update: address, order, services

+++ backend/internal/models/service.go
@@ -9,7 +9,7 @@ import (
 type Service struct {
 	ID              bson.ObjectID `json:"id" bson:"_id,omitempty"`
 	Name            string        `json:"name" bson:"name" validate:"required"`
-	Price           float64       `json:"price" bson:"price" validate:"required"`
+	Price           int           `json:"price" bson:"price" validate:"required"`
 	WorkersQuantity int           `json:"workers_quantity" bson:"workers_quantity" validate:"required"`
 	Description     string        `json:"description" bson:"description"`
 	Consumables     []Consumable  `json:"consumables,omitempty" bson:"consumables,omitempty" validate:"required"`
